server/requirements.txt
# Requirements for ReactGPT with Groq API
fastapi
uvicorn[standard]
python-dotenv
groq
pydantic
requests
